# public/swagger/v1/swagger.json
{
	"components": {
...
		"schemas": {
			"Link": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Meta": {
				"properties": {
					"publisher": {
						"type": "string"
					},
					"licence": {
						"type": "string"
					},
					"documentation": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"hasFormat": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"publisher",
					"licence",
					"documentation",
					"version",
					"comment",
					"hasFormat",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AuthorityBasic": {
				"properties": {
					"localAuthorityId": {
						"type": "number",
						"format": "double"
					},
					"localAuthorityIdCode": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"establishmentCount": {
						"type": "number",
						"format": "double"
					},
					"schemeType": {
						"type": "number",
						"format": "double"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"localAuthorityId",
					"localAuthorityIdCode",
					"name",
					"establishmentCount",
					"schemeType",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuthorityDetailBasicPaginationGetResponse": {
				"properties": {
					"authorities": {
						"items": {
							"$ref": "#/components/schemas/AuthorityBasic"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"authorities",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Region": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetaRegion": {
				"properties": {
					"dataSource": {
						"type": "string"
					},
					"extractDate": {
						"type": "string",
						"format": "date-time"
					},
					"itemCount": {
						"type": "number",
						"format": "double"
					},
					"returncode": {
						"type": "string"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"dataSource",
					"extractDate",
					"itemCount",
					"returncode",
					"totalCount",
					"totalPages",
					"pageSize",
					"pageNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRegionDetailPaginationGetResponse": {
				"properties": {
					"regions": {
						"items": {
							"$ref": "#/components/schemas/Region"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/MetaRegion"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"regions",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRegionDetailBasicPaginationGetResponse": {
				"properties": {
					"regions": {
						"items": {
							"$ref": "#/components/schemas/Region"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/MetaRegion"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"regions",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IRegionDetailByIdGetResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"name": "authorization",
				"in": "header",
				"bearerFormat": "JWT"
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY"
			}
		}
	},
	"info": {
		"title": "Food Standards Agency API",
		"version": "v1",
		"description": "The FSA Food Alerts API provides access to current and recent Food Alerts: Allergy Alerts (AA), Product Recall Information Notices (PRIN) and Food Alerts for Action (FAFA). It provides applications with the facility to list alerts matching some filter criterion, and to retrieve a description of an alert.",
		"contact": {
			"name": "Oraclo LTD",
			"email": "admin@orcalo.co.uk",
			"url": "https://orcalo.co.uk/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/v1/authorities/basic/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetRegionDetailBasicPagination",
				"responses": {
					"200": {
						"description": "IAuthorityDetailBasicPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAuthorityDetailBasicPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a basic list of authorities (id/name), details can be retrieved by calling authorities/{id}, page parameters allow for page number and size specification.",
				"tags": [
					"Authorities"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/regions/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetRegionDetailPagination",
				"responses": {
					"200": {
						"description": "IRegionDetailPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRegionDetailPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of all regions, page parameters allow for page number and size specification.",
				"tags": [
					"Regions"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/regions/basic/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetRegionDetailBasicPagination",
				"responses": {
					"200": {
						"description": "IRegionDetailBasicPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRegionDetailBasicPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a basic list of regions, details can be retrieved by calling regions{id}, page parameters allow for page number and size specification.",
				"tags": [
					"Regions"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/regions/{id}": {
			"get": {
				"operationId": "GetRegionDetailById",
				"responses": {
					"200": {
						"description": "IRegionDetailByIdGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRegionDetailByIdGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of a single region, selected by Id.",
				"tags": [
					"Regions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}